third_party/blink/web_tests/virtual/text-antialias/text-container-bounding-rect-expected.txt

Bug 454760: getBoundingClientRect() on a container may be smaller than getBoundingClientRect() on the text it contains.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS divrect.width is >= spanrect.width
Some text
